Refactor iris platform layout
