+++ lib/iris/src/iris/cluster/platform/manual.py
@@ -27,23 +27,34 @@ Unlike cloud platforms, manual hosts are:
 from __future__ import annotations
 
 import logging
+import shlex
 from collections.abc import Callable
 from concurrent.futures import ThreadPoolExecutor, as_completed
+from contextlib import AbstractContextManager, nullcontext
 
+from iris.cluster.platform.base import VmBootstrapSpec, VmInfo, VmState
+from iris.cluster.platform.bootstrap import (
+    CONTROLLER_CONTAINER_NAME,
+    CONTROLLER_STOP_SCRIPT,
+    DEFAULT_CONTROLLER_PORT,
+    wait_healthy_via_ssh,
+)
+from iris.cluster.platform.ssh import DirectSshConnection, run_streaming_with_retry
 from iris.cluster.platform.worker_vm import (
     WorkerVm,
     PoolExhaustedError,
     SshConfig,
     VmFactory,
     VmRegistry,
+    TrackedVmFactory,
 )
 from iris.cluster.platform.vm_platform import (
     MAX_RECONCILE_WORKERS,
     VmGroupProtocol,
     VmGroupStatus,
     VmSnapshot,
+    VmManagerProtocol,
 )
-from iris.cluster.platform.ssh import DirectSshConnection
 from iris.rpc import config_pb2, vm_pb2
 from iris.time_utils import Duration, Timestamp
 
@@ -377,3 +388,198 @@ echo "[iris-init] Using static controller at $CONTROLLER_ADDRESS"
     def stop(self) -> None:
         """No-op: ManualVmManager has no background threads to stop."""
         pass
+
+
+class _ManualPlatformOps:
+    def __init__(self, ssh_config: SshConfig, bootstrap: config_pb2.BootstrapConfig):
+        self._ssh_config = ssh_config
+        self._bootstrap = bootstrap
+
+    def list_slices(self, group_config: config_pb2.ScaleGroupConfig, *, zone: str | None = None) -> list[str]:
+        hosts = list(group_config.manual.hosts)
+        if not hosts:
+            return []
+        port = self._bootstrap.worker_port or 10001
+        running: list[str] = []
+        for host in hosts:
+            conn = _direct_ssh(_apply_manual_overrides(self._ssh_config, group_config.manual), host)
+            try:
+                result = conn.run(
+                    f"curl -sf http://localhost:{port}/health",
+                    timeout=Duration.from_seconds(10),
+                )
+                if result.returncode == 0:
+                    running.append(host)
+            except Exception as exc:
+                logger.warning("Manual slice probe failed for host %s: %s", host, exc)
+                continue
+        return running
+
+    def delete_slice(
+        self,
+        group_config: config_pb2.ScaleGroupConfig,
+        slice_id: str,
+        *,
+        zone: str | None = None,
+    ) -> None:
+        conn = _direct_ssh(_apply_manual_overrides(self._ssh_config, group_config.manual), slice_id)
+        cmd = (
+            "sudo docker stop iris-worker 2>/dev/null || "
+            "sudo docker kill iris-worker 2>/dev/null || true; "
+            "sudo docker rm -f iris-worker 2>/dev/null || true"
+        )
+        conn.run(cmd, timeout=Duration.from_seconds(60))
+
+
+class ManualPlatform:
+    def __init__(
+        self,
+        label_prefix: str,
+        bootstrap_config: config_pb2.BootstrapConfig,
+        timeout_config: config_pb2.TimeoutConfig,
+        ssh_config: config_pb2.SshConfig,
+    ):
+        """Create manual platform with explicit config sections."""
+        self._bootstrap = bootstrap_config
+        self._timeouts = timeout_config
+        self._label_prefix = label_prefix
+        self._ssh = _ssh_config_to_dataclass(ssh_config)
+
+    def vm_ops(self) -> _ManualPlatformOps:
+        return _ManualPlatformOps(self._ssh, self._bootstrap)
+
+    def tunnel(
+        self,
+        controller_address: str,
+        local_port: int | None = None,
+        timeout: float | None = None,
+        tunnel_logger: logging.Logger | None = None,
+    ) -> AbstractContextManager[str]:
+        return nullcontext(controller_address)
+
+    def vm_manager(
+        self,
+        group_config: config_pb2.ScaleGroupConfig,
+        vm_factory: TrackedVmFactory,
+        *,
+        dry_run: bool = False,
+    ) -> VmManagerProtocol:
+        if group_config.vm_type != config_pb2.VM_TYPE_MANUAL_VM:
+            raise ValueError(f"Unsupported vm_type for manual platform: {group_config.vm_type}")
+        hosts = list(group_config.manual.hosts)
+        if not hosts:
+            raise ValueError(f"Manual scale group {group_config.name} missing hosts")
+        ssh_config = _apply_manual_overrides(self._ssh, group_config.manual)
+        return ManualVmManager(
+            hosts=hosts,
+            config=group_config,
+            bootstrap_config=self._bootstrap,
+            timeouts=self._timeouts,
+            vm_factory=vm_factory,
+            ssh_config=ssh_config,
+            label_prefix=self._label_prefix,
+            dry_run=dry_run,
+        )
+
+    def list_vms(self, *, tag: str | None = None, zone: str | None = None) -> list[VmInfo]:
+        return []
+
+    def start_vms(self, spec: VmBootstrapSpec, *, zone: str | None = None) -> list[VmInfo]:
+        if spec.role != "controller":
+            raise NotImplementedError("ManualPlatform.start_vms currently supports controller role only")
+
+        host = spec.provider_overrides.get("host")
+        if not host:
+            raise ValueError("provider_overrides.host is required for manual controller bootstrap")
+
+        port = spec.provider_overrides.get("port", DEFAULT_CONTROLLER_PORT)
+        ssh_user = spec.provider_overrides.get("ssh_user", self._ssh.user)
+        ssh_key_file = spec.provider_overrides.get("ssh_key_file", self._ssh.key_file)
+
+        conn = DirectSshConnection(
+            host=host,
+            user=ssh_user,
+            port=self._ssh.port,
+            key_file=ssh_key_file or None,
+            connect_timeout=self._ssh.connect_timeout,
+        )
+
+        if spec.bootstrap_script is None:
+            raise ValueError("bootstrap_script is required to start controller VM")
+
+        def on_line(line: str) -> None:
+            logger.info("[%s] %s", host, line)
+
+        run_streaming_with_retry(
+            conn,
+            f"bash -c {shlex.quote(spec.bootstrap_script)}",
+            on_line=on_line,
+        )
+
+        if port and not wait_healthy_via_ssh(conn, port):
+            raise RuntimeError(f"Controller at {host}:{port} failed health check after bootstrap")
+
+        address = spec.provider_overrides.get("address", f"http://{host}:{port}")
+
+        return [
+            VmInfo(
+                vm_id=host,
+                address=address,
+                zone=None,
+                labels=dict(spec.labels),
+                state=VmState.RUNNING,
+                created_at_ms=Timestamp.now().epoch_ms(),
+            )
+        ]
+
+    def stop_vms(self, ids: list[str], *, zone: str | None = None) -> None:
+        for host in ids:
+            conn = DirectSshConnection(
+                host=host,
+                user=self._ssh.user,
+                port=self._ssh.port,
+                key_file=self._ssh.key_file,
+                connect_timeout=self._ssh.connect_timeout,
+            )
+            try:
+                conn.run(
+                    f"bash -c {shlex.quote(CONTROLLER_STOP_SCRIPT.format(container_name=CONTROLLER_CONTAINER_NAME))}",
+                    timeout=Duration.from_seconds(60),
+                )
+            except Exception as exc:
+                logger.warning("Failed to stop controller on %s: %s", host, exc)
+
+
+def _direct_ssh(ssh_config: SshConfig, host: str) -> DirectSshConnection:
+    return DirectSshConnection(
+        host=host,
+        user=ssh_config.user,
+        port=ssh_config.port,
+        key_file=ssh_config.key_file or None,
+        connect_timeout=ssh_config.connect_timeout,
+    )
+
+
+def _ssh_config_to_dataclass(ssh: config_pb2.SshConfig) -> SshConfig:
+    """Convert proto SshConfig to dataclass SshConfig."""
+    from iris.config import DEFAULT_CONFIG, DEFAULT_SSH_PORT
+
+    connect_timeout = (
+        Duration.from_proto(ssh.connect_timeout)
+        if ssh.HasField("connect_timeout") and ssh.connect_timeout.milliseconds > 0
+        else Duration.from_proto(DEFAULT_CONFIG.ssh.connect_timeout)
+    )
+    return SshConfig(
+        user=ssh.user or DEFAULT_CONFIG.ssh.user,
+        key_file=ssh.key_file or None,
+        port=DEFAULT_SSH_PORT,
+        connect_timeout=connect_timeout,
+    )
+
+
+def _apply_manual_overrides(ssh_config: SshConfig, manual: config_pb2.ManualProvider) -> SshConfig:
+    return SshConfig(
+        user=manual.ssh_user or ssh_config.user,
+        key_file=manual.ssh_key_file or ssh_config.key_file,
+        connect_timeout=ssh_config.connect_timeout,
+    )
