+++ lib/iris/src/iris/cluster/platform/worker_vm.py
@@ -26,11 +26,11 @@ This module provides:
 from __future__ import annotations
 
 import logging
-import shlex
 import threading
 from dataclasses import dataclass, field
 from typing import Protocol
 
+from iris.cluster.platform.bootstrap import build_worker_bootstrap_script
 from iris.cluster.platform.ssh import (
     SshConnection,
     run_streaming_with_retry,
@@ -58,218 +58,6 @@ class BootstrapError(Exception):
 
 class QuotaExceededError(Exception):
     """Raised when cloud provider quota is exceeded."""
-
-
-# ============================================================================
-# Bootstrap Script
-# ============================================================================
-
-# Bootstrap script template for worker VMs.
-#
-# This script expects CONTROLLER_ADDRESS to be set as a shell variable before
-# the main script runs. Platforms inject a discovery preamble that sets this:
-# - GCP (TpuVmManager): Queries GCP instance metadata
-# - Manual (ManualVmManager): Uses static address from config
-BOOTSTRAP_SCRIPT = """
-set -e
-
-echo "[iris-init] Starting Iris worker bootstrap"
-
-# Fetch TPU metadata from GCE and export as environment variables.
-# GCP TPU VMs expose TPU info via instance metadata, not environment variables.
-echo "[iris-init] Probing TPU metadata..."
-METADATA_URL="http://metadata.google.internal/computeMetadata/v1/instance/attributes"
-METADATA_HEADER="Metadata-Flavor: Google"
-
-# Derive TPU slice name from instance name by stripping worker suffix.
-# Handles both -wN (original) and -w-N (GCP multi-host) formats.
-# Example: "iris-v5litepod_16-abc123-w0" -> "iris-v5litepod_16-abc123"
-# Example: "t1v-n-598bede5-w-0" -> "t1v-n-598bede5"
-INSTANCE_NAME=$(curl -sf -H "$METADATA_HEADER" \
-    "http://metadata.google.internal/computeMetadata/v1/instance/name" 2>/dev/null || echo "")
-if [ -n "$INSTANCE_NAME" ]; then
-    export TPU_NAME=$(echo "$INSTANCE_NAME" | sed 's/-w-*[0-9]*$//')
-fi
-
-# Fetch accelerator-type as TPU_TYPE (e.g., v5litepod-16)
-export TPU_TYPE=$(curl -sf -H "$METADATA_HEADER" "$METADATA_URL/accelerator-type" 2>/dev/null || echo "")
-
-# Fetch agent-worker-number as TPU_WORKER_ID
-export TPU_WORKER_ID=$(curl -sf -H "$METADATA_HEADER" "$METADATA_URL/agent-worker-number" 2>/dev/null || echo "")
-
-# Fetch worker hostnames for multi-host slices (JSON array of network endpoints)
-TPU_HOSTNAMES_RAW=$(curl -sf -H "$METADATA_HEADER" "$METADATA_URL/worker-network-endpoints" 2>/dev/null || echo "")
-if [ -n "$TPU_HOSTNAMES_RAW" ]; then
-    # Format is "unknown:unknown:ip1,unknown:unknown:ip2,..." — extract IPs (fields with dots)
-    export TPU_WORKER_HOSTNAMES=$(echo "$TPU_HOSTNAMES_RAW" | \
-        tr ',' '\\n' | while read -r entry; do echo "$entry" | tr ':' '\\n' | grep '[.]'; done | \\
-        paste -sd ',' - 2>/dev/null || echo "")
-fi
-
-# Fetch chips per host bounds from tpu-env metadata (YAML-like key: value format)
-TPU_ENV_RAW=$(curl -sf -H "$METADATA_HEADER" "$METADATA_URL/tpu-env" 2>/dev/null || echo "")
-if [ -n "$TPU_ENV_RAW" ]; then
-    TOPO_RAW=$(echo "$TPU_ENV_RAW" | grep "^CHIPS_PER_HOST_BOUNDS:" | sed "s/.*: *'\\(.*\\)'/\\1/" | tr -d "'")
-    if [ -n "$TOPO_RAW" ]; then
-        export TPU_CHIPS_PER_HOST_BOUNDS="$TOPO_RAW"
-    fi
-fi
-
-if [ -n "$TPU_NAME" ]; then
-    echo "[iris-init] TPU detected: name=$TPU_NAME type=$TPU_TYPE worker_id=$TPU_WORKER_ID"
-else
-    echo "[iris-init] No TPU metadata found (this may be a non-TPU VM)"
-fi
-
-echo "[iris-init] Phase: prerequisites"
-
-# Install Docker if missing
-if ! command -v docker &> /dev/null; then
-    echo "[iris-init] Installing Docker..."
-    curl -fsSL https://get.docker.com | sudo sh
-    sudo systemctl enable docker
-    sudo systemctl start docker
-    echo "[iris-init] Docker installed"
-else
-    echo "[iris-init] Docker already installed"
-fi
-
-# Ensure docker daemon is running
-sudo systemctl start docker || true
-
-# Create cache directory
-sudo mkdir -p {cache_dir}
-
-echo "[iris-init] Phase: docker_pull"
-echo "[iris-init] Configuring docker authentication"
-# Configure docker for common GCP Artifact Registry regions
-sudo gcloud auth configure-docker \
-  europe-west4-docker.pkg.dev,us-central1-docker.pkg.dev,us-docker.pkg.dev --quiet 2>/dev/null || true
-
-echo "[iris-init] Pulling image: {docker_image}"
-sudo docker pull {docker_image}
-
-echo "[iris-init] Phase: worker_start"
-
-# Force-remove existing worker (handles restart policy race)
-sudo docker rm -f iris-worker 2>/dev/null || true
-
-# Clean up ALL iris-managed task containers by label
-echo "[iris-init] Cleaning up iris task containers"
-IRIS_CONTAINERS=$(sudo docker ps -aq --filter "label=iris.managed=true" 2>/dev/null || true)
-if [ -n "$IRIS_CONTAINERS" ]; then
-    sudo docker rm -f $IRIS_CONTAINERS 2>/dev/null || true
-fi
-
-# Start worker container without restart policy first (fail fast during bootstrap)
-# Note: CONTROLLER_ADDRESS is set by the discovery preamble prepended to this script
-sudo docker run -d --name iris-worker \
-    --network=host \
-    -v {cache_dir}:{cache_dir} \
-    -v /var/run/docker.sock:/var/run/docker.sock \
-    {env_flags} \
-    {docker_image} \
-    .venv/bin/python -m iris.cluster.worker.main serve \
-        --host 0.0.0.0 --port {worker_port} \
-        --controller-address "$CONTROLLER_ADDRESS"
-
-echo "[iris-init] Worker container started"
-echo "[iris-init] Phase: registration"
-echo "[iris-init] Waiting for worker to register with controller..."
-
-# Wait for worker to be healthy (poll health endpoint)
-for i in $(seq 1 60); do
-    # Check if container is still running
-    if ! sudo docker ps -q -f name=iris-worker | grep -q .; then
-        echo "[iris-init] ERROR: Worker container exited unexpectedly"
-        echo "[iris-init] Container status:"
-        sudo docker ps -a -f name=iris-worker --format "table {{{{.Status}}}}\t{{{{.State}}}}"
-        echo "[iris-init] Container logs:"
-        sudo docker logs iris-worker --tail 100
-        exit 1
-    fi
-
-    if curl -sf http://localhost:{worker_port}/health > /dev/null 2>&1; then
-        echo "[iris-init] Worker is healthy"
-        # Now add restart policy for production
-        sudo docker update --restart=unless-stopped iris-worker
-        echo "[iris-init] Bootstrap complete"
-        exit 0
-    fi
-    sleep 2
-done
-
-echo "[iris-init] ERROR: Worker failed to become healthy after 120s"
-echo "[iris-init] Container status:"
-sudo docker ps -a -f name=iris-worker --format "table {{{{.Status}}}}\t{{{{.State}}}}"
-echo "[iris-init] Container logs:"
-sudo docker logs iris-worker --tail 100
-exit 1
-"""
-
-
-def _build_env_flags(config: config_pb2.BootstrapConfig, vm_address: str) -> str:
-    """Generate docker -e flags with proper escaping.
-
-    Note: IRIS_CONTROLLER_ADDRESS is set from the shell variable $CONTROLLER_ADDRESS
-    which is populated by the discovery preamble.
-
-    TPU environment variables (TPU_NAME, TPU_WORKER_ID, etc.) are passed through
-    from the host if they exist. These are set by GCP on TPU VMs and are required
-    for the worker to register with tpu-name and tpu-worker-id attributes needed
-    for coscheduled job scheduling.
-    """
-    flags = []
-    for k, v in config.env_vars.items():
-        flags.append(f"-e {shlex.quote(k)}={shlex.quote(v)}")
-    # Use shell variable set by discovery preamble, not Python string
-    flags.append('-e IRIS_CONTROLLER_ADDRESS="$CONTROLLER_ADDRESS"')
-    # Inject VM address so worker can include it in registration for autoscaler tracking
-    if vm_address:
-        flags.append(f"-e IRIS_VM_ADDRESS={shlex.quote(vm_address)}")
-
-    # Pass through TPU environment variables from host if they exist.
-    # These are fetched from GCE metadata by the bootstrap script preamble.
-    # TPU_NAME = instance name with -w{{N}} suffix stripped (TPU slice name) for coscheduling group_by
-    # TPU_TYPE = accelerator-type (e.g., v5litepod-16) for topology lookup
-    tpu_env_vars = [
-        "TPU_NAME",
-        "TPU_TYPE",
-        "TPU_WORKER_ID",
-        "TPU_WORKER_HOSTNAMES",
-        "TPU_CHIPS_PER_HOST_BOUNDS",
-    ]
-    for var in tpu_env_vars:
-        # Use shell syntax to pass through only if set on host
-        flags.append(f'-e {var}="${{{var}:-}}"')
-
-    return " ".join(flags)
-
-
-def _build_bootstrap_script(
-    config: config_pb2.BootstrapConfig,
-    vm_address: str,
-    discovery_preamble: str = "",
-) -> str:
-    """Build the bootstrap script from config.
-
-    Args:
-        config: Bootstrap configuration
-        vm_address: VM IP address for autoscaler tracking
-        discovery_preamble: Shell script fragment that sets CONTROLLER_ADDRESS.
-            This is prepended to the main bootstrap script and must define
-            the CONTROLLER_ADDRESS variable for the worker to connect.
-    """
-    env_flags = _build_env_flags(config, vm_address)
-    main_script = BOOTSTRAP_SCRIPT.format(
-        cache_dir=config.cache_dir or "/var/cache/iris",
-        docker_image=config.docker_image,
-        worker_port=config.worker_port or 10001,
-        env_flags=env_flags,
-    )
-    if discovery_preamble:
-        return discovery_preamble + main_script
-    return main_script
 
 
 # ============================================================================
@@ -384,7 +172,7 @@ class WorkerVm:
 
             # Run bootstrap - pass vm_address for IRIS_VM_ADDRESS env var
             logger.info("VM %s: Starting bootstrap", vm_id)
-            script = _build_bootstrap_script(self._bootstrap_config, vm_address, self._discovery_preamble)
+            script = build_worker_bootstrap_script(self._bootstrap_config, vm_address, self._discovery_preamble)
 
             result = run_streaming_with_retry(
                 self._conn,
@@ -515,7 +303,7 @@ class WorkerVm:
         self._transition(vm_pb2.VM_STATE_INITIALIZING)
 
         # Build and run bootstrap script
-        script = _build_bootstrap_script(
+        script = build_worker_bootstrap_script(
             self._bootstrap_config,
             self.info.address,
             self._discovery_preamble,
